move ContactUsForm inside of ContactUsModal
and have default email value from firebase
