--- new-nextjs-client/src/components/LeftSideBar/ContactUsModal/ContactUsModal.module.scss
@@ -1,8 +1,0 @@
-.logoContainer {
-  position: relative;
-  height: 5.5rem;
-  display: flex;
-  justify-self: center;
-  margin-bottom: 1rem;
-  margin-top: 3rem;
-}
